requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
SQLAlchemy==2.0.40
python-dotenv==1.1.0
jinja2==3.1.6
openai==1.75.0
